Module Account.Bootstrap

val alias : int -> string

Standard name for a bootstrap account parameterised by an integer. This alias can be used to name new bootstrap accounts.

val keys : key array

The default bootstrap keys.