Module Client.Admin

Run tezos-admin-client commands.

val trust_address : ?node:Node.t -> peer:Node.t -> t -> unit Lwt.t

Ask a node to trust the address and port of another node.

val spawn_trust_address : ?node:Node.t -> peer:Node.t -> t -> Tezt.Process.t

Same as trust_address, but do not wait for the process to exit.

val connect_address : ?node:Node.t -> peer:Node.t -> t -> unit Lwt.t

Connect a node to another peer.

val spawn_connect_address : ?node:Node.t -> peer:Node.t -> t -> Tezt.Process.t

Same as connect_address, but do not wait for the process to exit.

val kick_peer : ?node:Node.t -> peer:string -> t -> unit Lwt.t

Kick a peer.

peer is the identity of the peer to kick. You can get it with Node.wait_for_identity for instance.

val spawn_kick_peer : ?node:Node.t -> peer:string -> t -> Tezt.Process.t

Same as kick_peer, but do not wait for the process to exit.