Module Client.Sc_rollup

val remember_smart_rollup : ?hooks:Tezt_wrapper.Process.hooks -> ?force:bool -> t -> alias:string -> address:string -> unit Runnable.process

Run octez-client remember smart rollup <alias> <address>.

val list_known_smart_rollups : ?hooks:Tezt_wrapper.Process.hooks -> t -> (string * string) list Runnable.process

Run octez-client list known smart rollups.

val forget_all_smart_rollups : ?hooks:Tezt_wrapper.Process.hooks -> ?force:bool -> t -> unit Runnable.process

Run octez-client forget all smart rollups.

val show_known_smart_rollup : ?hooks:Tezt_wrapper.Process.hooks -> t -> alias:string -> string Runnable.process

Run octez-client show known smart rollup <alias>.

val originate : ?hooks:Tezt_wrapper.Process.hooks -> ?wait:string -> ?force:bool -> ?burn_cap:Tez.t -> ?whitelist:string list -> alias:string -> src:string -> kind:string -> parameters_ty:string -> boot_sector:string -> t -> string Lwt.t

Run octez-client originate sc rollup <alias> from <src> of kind <kind> booting with <boot_sector>.

val spawn_originate : ?hooks:Tezt_wrapper.Process.hooks -> ?wait:string -> ?force:bool -> ?burn_cap:Tez.t -> ?whitelist:string list -> alias:string -> src:string -> kind:string -> parameters_ty:string -> boot_sector:string -> t -> Tezt_wrapper.Process.t

Same as originate, but do not wait for the process to exit.

val send_message : ?hooks:Tezt_wrapper.Process.hooks -> ?wait:string -> ?burn_cap:Tez.t -> ?fee:Tez.t -> ?fee_cap:Tez.t -> msg:string -> src:string -> t -> unit Lwt.t

Run octez-client send rollup message <msg> from <src>.

val spawn_send_message : ?hooks:Tezt_wrapper.Process.hooks -> ?wait:string -> ?burn_cap:Tez.t -> ?fee:Tez.t -> ?fee_cap:Tez.t -> msg:string -> src:string -> t -> Tezt_wrapper.Process.t

Same as send_message, but do not wait for the process to exit.

val publish_commitment : ?hooks:Tezt_wrapper.Process.hooks -> ?wait:string -> ?burn_cap:Tez.t -> src:string -> sc_rollup:string -> compressed_state:string -> inbox_level:int -> predecessor:string -> number_of_ticks:int -> t -> unit Runnable.process

Run octez-client publish commitment from <src> for sc rollup <sc_rollup> with compressed state <compressed_state> at inbox level <inbox_level> and predecessor <predecessor> and number of ticks <number_of_ticks>.

val cement_commitment : Protocol.t -> ?hooks:Tezt_wrapper.Process.hooks -> ?wait:string -> ?burn_cap:Tez.t -> hash:string -> src:string -> dst:string -> t -> unit Runnable.process

Run octez-client cement commitment <hash> from <src> for sc rollup <rollup>.

val timeout : ?expect_failure:bool -> ?hooks:Tezt_wrapper.Process.hooks -> ?wait:string -> ?burn_cap:Tez.t -> staker1:string -> staker2:string -> src:string -> dst:string -> t -> unit Runnable.process

Run octez-client timeout dispute on sc rollup <dst> with <staker1> against <staker2> from <src>.

val submit_recover_bond : ?wait:string -> ?burn_cap:Tez.t -> ?storage_limit:int -> ?fee:Tez.t -> ?hooks:Tezt_wrapper.Process.hooks -> rollup:string -> src:string -> staker:string -> t -> unit Runnable.process

Run octez-client submit sc rollup recover bond of <staker> for <sc_rollup> from <src>.

val execute_outbox_message : ?wait:string -> ?burn_cap:Tez.t -> ?storage_limit:int -> ?fee:Tez.t -> ?hooks:Tezt_wrapper.Process.hooks -> rollup:string -> src:string -> commitment_hash:string -> proof:string -> t -> unit Runnable.process

Run octez-client execute outbox message of sc rollup <rollup> from <src> for commitment hash <hash> and output proof <proof>.