Module Tezt_tezos.Cluster

Set up networks involving large amount of nodes.

The applications of this module range from running small cliques to complex network topologies involving as many nodes as your computer can handle.

The general approach is to create some nodes, then declare connections using functions such as connect, clique, ring or star, and finally start all nodes with start.

For instance, here is how to create a star topology, with a center node connected to 5 other nodes: let center = Node.create [] in let other_star_nodes = Cluster.create 5 [] in Cluster.star center other_star_nodes; Here is how to connect the center node of this star to a clique of 4 other nodes: let clique_nodes = Cluster.create 4 [] in Cluster.clique clique_nodes; Cluster.connect [center] clique_nodes; Finally, here is how to start the network: Cluster.start (clique_nodes @ center :: other_star_nodes)

Batch Node Creation

val create : ?path:string -> ?name:string -> int -> Node.argument list -> Node.t list

Create several nodes.

Usage: create count arguments

Create count nodes (using Node.create) sharing the same arguments. Each node is named name ^ "." ^ i where i is the index of the node in the resulting list, starting from 1.

Topologies

val connect : Node.t list -> Node.t list -> unit

Connect all nodes of a cluster to all nodes of another cluster.

Usage: connect a b

Add all nodes of b as --peer arguments to all nodes of a. This is a generalization of Node.add_peer for lists.

val clique : Node.t list -> unit

Connect all nodes of a cluster together.

For a cluster node1; node2; ...; nodeN, this adds:

node2; ...; nodeN as --peers to node1;

node3; ...; nodeN as --peers to node2;

etc.

val ring : Node.t list -> unit

Connect nodes to form a ring.

For a cluster node1; node2; ...; nodeN, this adds:

node2 as --peer to node1;

node3 as --peer to node2;

...

and finally node1 as --peer to nodeN.

val star : Node.t -> Node.t list -> unit

Connect nodes to form a star.

Usage: star center other_nodes

Add all other_nodes as --peers to center.

Meta-Topologies

The following functions are similar to the above functions that create topologies, except that they connect clusters using connect instead of nodes using Node.add_peer.

For instance, a meta-clique a; b; c connects:

all nodes of a with all nodes of b and c;

all nodes of b with all nodes of c. The result itself is not necessarily a clique since nodes of each sub-cluster are not necessarily connected between themselves.

val meta_clique : Node.t list list -> unit

Meta-version of clique.

val meta_ring : Node.t list list -> unit

Meta-version of ring.

val meta_star : Node.t list -> Node.t list list -> unit

Meta-version of star.

Running

val start : Node.t list -> unit Lwt.t

Start a cluster of nodes.

This runs Node.identity_generate, Node.config_init and Node.run for all nodes. This is similar to iterating Node.init except that the node are already created.