Module Tezt_tezos.Codec

val encode : ?⁠path:string -> ?⁠hooks:Tezt.Process.hooks -> name:string -> Tezt.JSON.u -> string Lwt.t

Run tezos-codec encode <name> from <json>.

val spawn_encode : ?⁠path:string -> ?⁠hooks:Tezt.Process.hooks -> name:string -> Tezt.JSON.u -> Tezt.Process.t

Same as encode, but do not wait for the process to exit.

val decode : ?⁠path:string -> ?⁠hooks:Tezt.Process.hooks -> name:string -> string -> Tezt.JSON.t Lwt.t

Run tezos-codec decode <name> from <binary>.

val spawn_decode : ?⁠path:string -> ?⁠hooks:Tezt.Process.hooks -> name:string -> string -> Tezt.Process.t

Same as decode, but do not wait for the process to exit.

val dump_encodings : ?⁠path:string -> unit -> Tezt.JSON.t Lwt.t

Run tezos-codec dump encodings.

val spawn_dump_encodings : ?⁠path:string -> unit -> Tezt.Process.t

Same as dump_encodings, but do not wait for the process to exit.