Module Tezt_tezos.Constant

val octez_client : Tezt_wrapper.Uses.t

val octez_admin_client : Tezt_wrapper.Uses.t

val octez_node : Tezt_wrapper.Uses.t

val octez_codec : Tezt_wrapper.Uses.t

val octez_snoop : Tezt_wrapper.Uses.t

val octez_protocol_compiler : Tezt_wrapper.Uses.t

val octez_dal_node : Tezt_wrapper.Uses.t

val octez_dac_node : Tezt_wrapper.Uses.t

val octez_dac_client : Tezt_wrapper.Uses.t

val octez_smart_rollup_node : Tezt_wrapper.Uses.t

val octez_evm_node : Tezt_wrapper.Uses.t

val etherlink_governance_observer : Tezt_wrapper.Uses.t

val octez_dsn_node : Tezt_wrapper.Uses.t

val octez_signer : Tezt_wrapper.Uses.t

val octez_injector_server : Tezt_wrapper.Uses.t

val smart_rollup_installer : Tezt_wrapper.Uses.t

val _octez_smart_rollup_wasm_debugger : Tezt_wrapper.Uses.t

val teztale_archiver : Tezt_wrapper.Uses.t

val teztale_server : Tezt_wrapper.Uses.t

val yes_wallet : Tezt_wrapper.Uses.t

module WASM : sig ... end

val octez_agnostic_baker : Tezt_wrapper.Uses.t

val released_executables : string

val experimental_executables : string

val default_host : string

Default hostname to use for endpoints when no specific one is required.

val activator : Account.key

Key pair used to activate a protocol from genesis with --network sandbox. The public key is hard-coded in the node.

val bootstrap1 : Tezt_tezos__Account.key

val bootstrap2 : Tezt_tezos__Account.key

val bootstrap3 : Tezt_tezos__Account.key

val bootstrap4 : Tezt_tezos__Account.key

val bootstrap5 : Tezt_tezos__Account.key

val all_secret_keys : Account.key list

val implicit_account_burn : Tez.t

The default burn for an implicit account.

val max_op_ttl : int

The default time to live of an operation (in block)

val manager_operation_gas_cost : protocol:'a -> int

Constant gas cost required for every manager operation. Should match Michelson_v1_gas.Cost_of.manager_operation.

val sc_rollup_compressed_state : string

A valid base58 encoded compressed state hash.

val tz4_account : Account.key

A valid base58 encoded layer-2 account to be used to test transaction and smart contract rollups.

val wasm_echo_kernel_boot_sector : string

The `echo` kernel that is listed in the “Smart Optimistic Rollups” section of the reference manual.