Module Tezt_tezos.Curl

val get : ?runner:Tezt_wrapper.Runner.t -> ?args:string list -> string -> Tezt_wrapper.JSON.t Runnable.process

get url returns a runnable requesting url with curl.

The response is parsed and returned as JSON. Fails if curl is not found in path.

val get_raw : ?runner:Tezt_wrapper.Runner.t -> ?args:string list -> string -> string Runnable.process

Same as get but does not parse the returned value

val post : ?runner:Tezt_wrapper.Runner.t -> ?args:string list -> string -> Tezt_wrapper.JSON.t -> Tezt_wrapper.JSON.t Runnable.process

post url data returns a runnable posting data to url with curl.

The response is parsed and returned as JSON. Fails if curl is not found in path.

val post_raw : ?runner:Tezt_wrapper.Runner.t -> ?args:string list -> string -> Tezt_wrapper.JSON.t -> string Runnable.process

Same as post but does not parse the returned value

val put : ?runner:Tezt_wrapper.Runner.t -> ?args:string list -> string -> Tezt_wrapper.JSON.t -> Tezt_wrapper.JSON.t Runnable.process

put url data returns a runnable putting data to url with curl.

The response is parsed and returned as JSON. Fails if curl is not found in path.

val put_raw : ?runner:Tezt_wrapper.Runner.t -> ?args:string list -> string -> Tezt_wrapper.JSON.t -> string Runnable.process

Same as put but does not parse the returned value