Module Tezt_tezos.Dac_helper

Helpers for configuring DAC scenarios.

module Scenarios : sig ... end

val with_coordinator_node : ?name:string -> ?sc_rollup_node:Sc_rollup_node.t -> ?pvm_name:string -> ?wait_ready:bool -> ?allow_v1_api:bool -> committee_members:Account.key list -> Node.t -> Client.t -> (Dac_node.t -> Account.key list -> 'a Lwt.t) -> 'a Lwt.t

Initializes a a Coordinator Dac node. DAC Committee of size committee_size will be generated and pre-configured in the Coordfinator node. An additional custom_committee_members can be provided to configure fixed members of the committee (useful for testing).

val with_committee_member : ?name:string -> ?sc_rollup_node:Sc_rollup_node.t -> ?pvm_name:string -> ?wait_ready:bool -> ?allow_v1_api:bool -> committee_member:Account.key -> Node.t -> Dac_node.t -> Client.t -> (Dac_node.t -> Account.key -> 'a Lwt.t) -> 'a Lwt.t

Initializes a a Committee Member Dac node with key committee_member.

val with_observer : ?name:string -> ?sc_rollup_node:Sc_rollup_node.t -> ?pvm_name:string -> ?wait_ready:bool -> ?allow_v1_api:bool -> committee_member_rpcs:(string * int) list -> Node.t -> Dac_node.t -> Client.t -> (Dac_node.t -> 'b Lwt.t) -> 'b Lwt.t

Initializes a Observer Dac node with key committee_member.

val with_fresh_rollup : ?pvm_name:string -> ?hooks:Tezt_wrapper.Process_hooks.t -> Node.t -> Client.t -> string -> (string -> Sc_rollup_node.t -> 'a Lwt.t) -> 'a Lwt.t

Initializes a new Sc_rollup with empty boot sector, attached to an Sc_rollup_node.

val scenario_with_full_dac_infrastructure : ?supports:Protocol.supported_protocols -> ?tags:string list -> ?uses:(Protocol.t -> Tezt_wrapper.Uses.t list) -> ?pvm_name:string -> ?custom_committee_members:Account.key list -> ?commitment_period:int -> ?challenge_window:int -> ?event_sections_levels:(string * Daemon.Level.level) list -> ?node_arguments:Node.argument list -> ?allow_v1_api:bool -> ?allow_regression:bool -> __FILE__:string -> committee_size:int -> observers:int -> string -> (Scenarios.full -> unit Lwt.t) -> Protocol.t list -> unit

Initalizes a scenario with full DAC infrastruture. See Scenarios.full for components.

val scenario_with_layer1_node : ?tags:string list -> ?uses:(Protocol.t -> Tezt_wrapper.Uses.t list) -> ?commitment_period:int -> ?challenge_window:int -> ?event_sections_levels:(string * Daemon.Level.level) list -> ?node_arguments:Node.argument list -> __FILE__:string -> string -> (Protocol.t -> Node.t -> Client.t -> string -> unit Lwt.t) -> Protocol.t list -> unit

Initalizes a scenario with L1 node only.

module Call_endpoint : sig ... end

This module is syntactix sugar to call DAC RPC endpoints.