Module Dac_rpc.V0

V0 is experimental DAC API.

val api_prefix : string

api_prefix is a URL prefix of the V0 API endpoints.

val get_preimage : string -> string RPC_core.t

get_preimage hash requests the preimage of hash, consisting of a single page, from cctxt. When the request succeeds, the raw page will be returned as a sequence of bytes.

val make_put_dac_member_signature_request_body : dac_member_pkh:string -> root_hash:Hex.t -> Tezos_crypto.Aggregate_signature.signature -> Ezjsonm.value

make_put_dac_member_sigature_request_body ~dac_member_pkh ~root_hash signature creates a JSON body for "PUT v0/dac_member_signature".

val put_dac_member_signature : hex_root_hash:Hex.t -> dac_member_pkh:string -> signature:Tezos_crypto.Aggregate_signature.signature -> unit RPC_core.t

put_dac_member_signature hex_root_hash dac_member_pkh signature stores the signature generated from signing hex_root_hash by dac_member_pkh.

val get_missing_page : hex_root_hash:Hex.t -> string RPC_core.t

get_missing_page ~hex_root_hash calls "GET v0/missing_page/page_hash" endpoint.

val get_certificate : hex_root_hash:Hex.t -> (int * string * string * int) RPC_core.t

get_certificate ~hex_root_hash fetches the DAC certificate for the provided hex_root_hash.

val get_serialized_certificate : hex_root_hash:Hex.t -> string RPC_core.t

get_serialized_certificate ~hex_root_hash fetches the DAC certificate for the provided hex_root_hash with SDK kernel compatible root_hash encoding.

module Coordinator : sig ... end