Module V0.Coordinator

val post_preimage : payload:string -> string RPC_core.t

post_preimage ~payload sends a payload to the DAC Coordinator via a POST RPC call to v0/preimage. It returns a hex encoded root page hash, produced by Merkle_tree_V0 pagination scheme. On the backend side it also pushes root page hash of the preimage to all the subscribed DAC Members and Observers.