Module Dac_rpc.V1

V1 is a second major DAC API release which is currently work in progress.

val get_pages : string -> string RPC_core.t

get_pages hash requests the preimage of hash, consisting of a single page, from cctxt. When the request succeeds, the raw page will be returned as a sequence of bytes. This is achieved by calling "GET v1/pages".