Module Tezt_tezos.Mempool

type t = {

validated : string list;

branch_delayed : string list;

branch_refused : string list;

refused : string list;

outdated : string list;

unprocessed : string list;

}

The mempool type description.

val typ : t Tezt_wrapper.Check.typ

A comparable type for mempool where classification and ordering does not matter.

val classified_typ : t Tezt_wrapper.Check.typ

A comparable type for mempool where ordering does not matter.

val empty : t

val symmetric_diff : t -> t -> t

Symetric difference (union(a, b) - intersection(a, b))

val of_json : Tezt_wrapper.JSON.t -> t

Build a value of type t from a json returned by RPC.get_mempool_pending_operations.

val get_mempool : ?endpoint:Client.endpoint -> ?hooks:Tezt_wrapper.Process.hooks -> ?chain:string -> ?validated:bool -> ?branch_delayed:bool -> ?branch_refused:bool -> ?refused:bool -> ?outdated:bool -> ?validation_passes:int list -> ?sources:string list -> ?operation_hash:string list -> Client.t -> t Lwt.t

Call RPC.get_mempool_pending_operations and wrap the result in a value of type Mempool.t

val check_mempool : ?validated:string list -> ?branch_delayed:string list -> ?branch_refused:string list -> ?refused:string list -> ?outdated:string list -> ?unprocessed:string list -> t -> unit

Check that each field of t contains the same elements as the argument of the same name. Ordening does not matter. Omitted arguments default to the empty list. This is useful when we expect a sparse mempool.

module Config : sig ... end

Mempool filter configuration.