Module Node.Config_file

Node configuration files.

val read : t -> Tezt_wrapper.JSON.t Lwt.t

Read the configuration file (config.json) of a node.

val write : t -> Tezt_wrapper.JSON.t -> unit Lwt.t

Write the configuration file of a node, replacing the existing one.

val update : t -> (Tezt_wrapper.JSON.t -> Tezt_wrapper.JSON.t) -> unit Lwt.t

Update the configuration file of a node. If the node is already running, it needs to be restarted manually.

Example: Node.Config_file.update node (JSON.put ("p2p", new_p2p_config))

val set_sandbox_network_with_user_activated_upgrades : (int * Protocol.t) list -> Tezt_wrapper.JSON.t -> Tezt_wrapper.JSON.t

Set the network config to a sandbox with the given user activated upgrades.

val set_sandbox_network_with_user_activated_overrides : (string * string) list -> Tezt_wrapper.JSON.t -> Tezt_wrapper.JSON.t

Set the network config to a sandbox with the given user activated protocol overrides.

val set_sandbox_network_with_dal_config : Tezos_crypto_dal.Cryptobox.Config.t -> Tezt_wrapper.JSON.t -> Tezt_wrapper.JSON.t

Set the network config to a sandbox with the given dal config.

val update_network_with_user_activated_upgrades : (int * Protocol.t) list -> Tezt_wrapper.JSON.t -> Tezt_wrapper.JSON.t

Update the network config with the given user activated upgrades.

val set_prevalidator : ?operations_request_timeout:float -> ?max_refused_operations:int -> ?operations_batch_size:int -> Tezt_wrapper.JSON.t -> Tezt_wrapper.JSON.t

Set the prevalidator configuration in the given configuration.

val set_peer_validator : ?new_head_request_timeout:float -> Tezt_wrapper.JSON.t -> Tezt_wrapper.JSON.t

Set the peer_validator configuration in the given configuration.

val set_ghostnet_sandbox_network : ?user_activated_upgrades:(int * Protocol.t) list -> unit -> Tezt_wrapper.JSON.t -> Tezt_wrapper.JSON.t

Set the network config to a sandbox with the same chain_id than Ghostnet.

user_activated_upgrades can be given to add user-activated upgrades.