Module Tezt_tezos.Operation_receipt

val get_result_for : string -> ?check_previous:int -> Tezt_tezos__Client.t -> Tezt_wrapper.JSON.t list Lwt.t

Returns the content of the "operation_result" field from the metadata of the operation with hash operation in the check_previous blocks before the current head using the 'get_chain_block_operations' RPC.

val get_block_metadata : Tezt_tezos__Client.t -> Tezt_wrapper.JSON.t Lwt.t

module Balance_updates : sig ... end