Module Tezt_tezos.Protocol_compiler

Run octez-protocol-compiler commands.

In all functions below, path can be used to override the path to octez-protocol-compiler, which is Constant.tezos_protocol_compiler by default.

val compile : ?path:string -> ?hooks:Tezt.Process.hooks -> ?hash_only:bool -> string -> string Lwt.t

Run octez-protocol-compiler <protocol_dir> and return the hash of the compiled protocol.

If hash_only is set, then pass -hash-only to the protocol compiler.

val spawn_compile : ?path:string -> ?hooks:Tezt.Process.hooks -> ?hash_only:bool -> string -> Tezt.Process.t

Same as compile, but do not wait for the process to exit.