Module RPC.Contracts

Common protocol RPSs for contracts (i.e. under /contracts).

val get_all : ?node:Node.t -> ?hooks:Tezt.Process.hooks -> ?chain:string -> ?block:string -> Client.t -> string list Lwt.t

Call RPC /chain/chain/blocks/block/context/contracts

val get_all_delegates : ?node:Node.t -> ?hooks:Tezt.Process.hooks -> ?chain:string -> ?block:string -> Client.t -> string list Lwt.t

Call RPC /chain/chain/blocks/block/context/delegates

val spawn_get_all : ?node:Node.t -> ?hooks:Tezt.Process.hooks -> ?chain:string -> ?block:string -> Client.t -> Tezt.Process.t

Same as get_all, but do not wait for the process to exit.

val get : ?node:Node.t -> ?hooks:Tezt.Process.hooks -> ?chain:string -> ?block:string -> contract_id:string -> Client.t -> Tezt.JSON.t Lwt.t

Call RPC /chain/chain/blocks/block/context/contracts/contract_id

val spawn_get : ?node:Node.t -> ?hooks:Tezt.Process.hooks -> ?chain:string -> ?block:string -> contract_id:string -> Client.t -> Tezt.Process.t

Same as get, but do not wait for the process to exit.

val get_balance : ?node:Node.t -> ?hooks:Tezt.Process.hooks -> ?chain:string -> ?block:string -> contract_id:string -> Client.t -> Tezt.JSON.t Lwt.t

Call RPC /chain/chain/blocks/block/context/contracts/contract_id/balance

val spawn_get_balance : ?node:Node.t -> ?hooks:Tezt.Process.hooks -> ?chain:string -> ?block:string -> contract_id:string -> Client.t -> Tezt.Process.t

Same as get_balance, but do not wait for the process to exit.

val big_map_get : ?node:Node.t -> ?hooks:Tezt.Process.hooks -> ?chain:string -> ?block:string -> contract_id:string -> data:Tezt.JSON.u -> Client.t -> Tezt.JSON.t Lwt.t

Call RPC /chain/chain/blocks/block/context/contracts/contract_id/big_map_get

val spawn_big_map_get : ?node:Node.t -> ?hooks:Tezt.Process.hooks -> ?chain:string -> ?block:string -> contract_id:string -> data:Tezt.JSON.u -> Client.t -> Tezt.Process.t

Same as big_map_get, but do not wait for the process to exit.

val get_counter : ?node:Node.t -> ?hooks:Tezt.Process.hooks -> ?chain:string -> ?block:string -> contract_id:string -> Client.t -> Tezt.JSON.t Lwt.t

Call RPC /chain/chain/blocks/block/context/contracts/contract_id/counter

val spawn_get_counter : ?node:Node.t -> ?hooks:Tezt.Process.hooks -> ?chain:string -> ?block:string -> contract_id:string -> Client.t -> Tezt.Process.t

Same as get_counter, but do not wait for the process to exit.

val get_delegate : ?node:Node.t -> ?hooks:Tezt.Process.hooks -> ?chain:string -> ?block:string -> contract_id:string -> Client.t -> Tezt.JSON.t Lwt.t

Call RPC /chain/chain/blocks/block/context/contracts/contract_id/delegate

val spawn_get_delegate : ?node:Node.t -> ?hooks:Tezt.Process.hooks -> ?chain:string -> ?block:string -> contract_id:string -> Client.t -> Tezt.Process.t

Same as get_delegate, but do not wait for the process to exit.

val get_entrypoints : ?node:Node.t -> ?hooks:Tezt.Process.hooks -> ?chain:string -> ?block:string -> contract_id:string -> Client.t -> Tezt.JSON.t Lwt.t

Call RPC /chain/chain/blocks/block/context/contracts/contract_id/entrypoints

val spawn_get_entrypoints : ?node:Node.t -> ?hooks:Tezt.Process.hooks -> ?chain:string -> ?block:string -> contract_id:string -> Client.t -> Tezt.Process.t

Same as get_entrypoints, but do not wait for the process to exit.

val get_manager_key : ?node:Node.t -> ?hooks:Tezt.Process.hooks -> ?chain:string -> ?block:string -> contract_id:string -> Client.t -> Tezt.JSON.t Lwt.t

Call RPC /chain/chain/blocks/block/context/contracts/contract_id/manager_key

val spawn_get_manager_key : ?node:Node.t -> ?hooks:Tezt.Process.hooks -> ?chain:string -> ?block:string -> contract_id:string -> Client.t -> Tezt.Process.t

Same as get_manager_key, but do not wait for the process to exit.

val get_script : ?node:Node.t -> ?hooks:Tezt.Process.hooks -> ?chain:string -> ?block:string -> contract_id:string -> Client.t -> Tezt.JSON.t Lwt.t

Call RPC /chain/chain/blocks/block/context/contracts/contract_id/script

val spawn_get_script : ?node:Node.t -> ?hooks:Tezt.Process.hooks -> ?chain:string -> ?block:string -> contract_id:string -> Client.t -> Tezt.Process.t

Same as get_script, but do not wait for the process to exit.

val get_storage : ?node:Node.t -> ?hooks:Tezt.Process.hooks -> ?chain:string -> ?block:string -> contract_id:string -> Client.t -> Tezt.JSON.t Lwt.t

Call RPC /chain/chain/blocks/block/context/contracts/contract_id/storage

val spawn_get_storage : ?node:Node.t -> ?hooks:Tezt.Process.hooks -> ?chain:string -> ?block:string -> contract_id:string -> Client.t -> Tezt.Process.t

Same as get_storage, but do not wait for the process to exit.