Module RPC.Delegates

Common protocol RPSs for delegates (i.e. under /delegates).

val get_all : ?node:Node.t -> ?hooks:Tezt.Process.hooks -> ?chain:string -> ?block:string -> Client.t -> string list Lwt.t

Call RPC /chain/chain/blocks/block/context/delegates

val spawn_get_all : ?node:Node.t -> ?hooks:Tezt.Process.hooks -> ?chain:string -> ?block:string -> Client.t -> Tezt.Process.t

Same as get_all, but do not wait for the process to exit.

val get : ?node:Node.t -> ?hooks:Tezt.Process.hooks -> ?chain:string -> ?block:string -> pkh:string -> Client.t -> Tezt.JSON.t Lwt.t

Call RPC /chain/chain/blocks/block/context/delegates/pkh

val spawn_get : ?node:Node.t -> ?hooks:Tezt.Process.hooks -> ?chain:string -> ?block:string -> pkh:string -> Client.t -> Tezt.Process.t

Same as get, but do not wait for the process to exit.

val get_balance : ?node:Node.t -> ?hooks:Tezt.Process.hooks -> ?chain:string -> ?block:string -> pkh:string -> Client.t -> Tezt.JSON.t Lwt.t

Call RPC /chain/chain/blocks/block/context/delegates/pkh/balance

val spawn_get_balance : ?node:Node.t -> ?hooks:Tezt.Process.hooks -> ?chain:string -> ?block:string -> pkh:string -> Client.t -> Tezt.Process.t

Same as get_balance, but do not wait for the process to exit.

val get_deactivated : ?node:Node.t -> ?hooks:Tezt.Process.hooks -> ?chain:string -> ?block:string -> pkh:string -> Client.t -> Tezt.JSON.t Lwt.t

Call RPC /chain/chain/blocks/block/context/delegates/pkh/deactivated

val spawn_get_deactivated : ?node:Node.t -> ?hooks:Tezt.Process.hooks -> ?chain:string -> ?block:string -> pkh:string -> Client.t -> Tezt.Process.t

Same as get_deactivated, but do not wait for the process to exit.

val get_delegated_balance : ?node:Node.t -> ?hooks:Tezt.Process.hooks -> ?chain:string -> ?block:string -> pkh:string -> Client.t -> Tezt.JSON.t Lwt.t

Call RPC /chain/chain/blocks/block/context/delegates/pkh/delegated_balance

val spawn_get_delegated_balance : ?node:Node.t -> ?hooks:Tezt.Process.hooks -> ?chain:string -> ?block:string -> pkh:string -> Client.t -> Tezt.Process.t

Same as get_delegated_balance, but do not wait for the process to exit.

val get_delegated_contracts : ?node:Node.t -> ?hooks:Tezt.Process.hooks -> ?chain:string -> ?block:string -> pkh:string -> Client.t -> Tezt.JSON.t Lwt.t

Call RPC /chain/chain/blocks/block/context/delegates/pkh/delegated_contracts

val spawn_get_delegated_contracts : ?node:Node.t -> ?hooks:Tezt.Process.hooks -> ?chain:string -> ?block:string -> pkh:string -> Client.t -> Tezt.Process.t

Same as get_delegated_contracts, but do not wait for the process to exit.

val get_frozen_balance : ?node:Node.t -> ?hooks:Tezt.Process.hooks -> ?chain:string -> ?block:string -> pkh:string -> Client.t -> Tezt.JSON.t Lwt.t

Call RPC /chain/chain/blocks/block/context/delegates/pkh/frozen_balance

val spawn_get_frozen_balance : ?node:Node.t -> ?hooks:Tezt.Process.hooks -> ?chain:string -> ?block:string -> pkh:string -> Client.t -> Tezt.Process.t

Same as get_frozen_balance, but do not wait for the process to exit.

val get_frozen_balance_by_cycle : ?node:Node.t -> ?hooks:Tezt.Process.hooks -> ?chain:string -> ?block:string -> pkh:string -> Client.t -> Tezt.JSON.t Lwt.t

Call RPC /chain/chain/blocks/block/context/delegates/pkh/frozen_balance_by_cycle

val spawn_get_frozen_balance_by_cycle : ?node:Node.t -> ?hooks:Tezt.Process.hooks -> ?chain:string -> ?block:string -> pkh:string -> Client.t -> Tezt.Process.t

Same as get_frozen_balance_by_cycle, but do not wait for the process to exit.

val get_grace_period : ?node:Node.t -> ?hooks:Tezt.Process.hooks -> ?chain:string -> ?block:string -> pkh:string -> Client.t -> Tezt.JSON.t Lwt.t

Call RPC /chain/chain/blocks/block/context/delegates/pkh/grace_period

val spawn_get_grace_period : ?node:Node.t -> ?hooks:Tezt.Process.hooks -> ?chain:string -> ?block:string -> pkh:string -> Client.t -> Tezt.Process.t

Same as get_grace_period, but do not wait for the process to exit.

val get_staking_balance : ?node:Node.t -> ?hooks:Tezt.Process.hooks -> ?chain:string -> ?block:string -> pkh:string -> Client.t -> Tezt.JSON.t Lwt.t

Call RPC /chain/chain/blocks/block/context/delegates/pkh/staking_balance

val spawn_get_staking_balance : ?node:Node.t -> ?hooks:Tezt.Process.hooks -> ?chain:string -> ?block:string -> pkh:string -> Client.t -> Tezt.Process.t

Same as get_staking_balance, but do not wait for the process to exit.

val get_voting_power : ?node:Node.t -> ?hooks:Tezt.Process.hooks -> ?chain:string -> ?block:string -> pkh:string -> Client.t -> Tezt.JSON.t Lwt.t

Call RPC /chain/chain/blocks/block/context/delegates/pkh/voting_power

val spawn_get_voting_power : ?node:Node.t -> ?hooks:Tezt.Process.hooks -> ?chain:string -> ?block:string -> pkh:string -> Client.t -> Tezt.Process.t

Same as get_voting_power, but do not wait for the process to exit.