Module Runnable.Syntax

Part of the Runnable module that is intended to be opened.

type nonrec ('a, 'b) t = ('a, 'b) t = {

value : 'a;

run : 'a -> 'b Lwt.t;

}

Makes fields value and run available in the environment.

type ('a, 'b) runnable = ('a, 'b) t

val let*! : ('a, 'b) t -> ('b -> 'c Lwt.t) -> 'c Lwt.t

Same as run, then continue with the given function.

val let*? : ('a, 'b) t -> ('a -> 'c) -> 'c

Get the value of a runnable and pass it to a continuation.

You can also just access field value directly.