Module Sc_rollup_helpers.Installer_kernel_config

type move_args = {

from : string;

to_ : string;

}

Moves path from at path to_.

type reveal_args = {

hash : string;

to_ : string;

}

Reveals hash hash at path to_.

type set_args = {

value : string;

to_ : string;

}

Sets value value at path to_.

type instr =

| Move of move_args

| Reveal of reveal_args

| Set of set_args

type t = instr list

Set of instructions used by the installer-client.

val of_json : string -> t

of_json path parses the JSON file at path and returns the installer config. Note that the instruction Reveal cannot be expressed in JSON because of the YAML definition, therefore path must not contain these instructions.