Module Sc_rollup_node.Config_file

Sc node configuration files.

val filename : t -> string

Returns the configuration file name for a rollup node.

val read : t -> Tezt_wrapper.JSON.t

Read the configuration file (config.json) of an sc node.

val write : t -> Tezt_wrapper.JSON.t -> unit

Write the configuration file of an sc node, replacing the existing one.

val update : t -> (Tezt_wrapper.JSON.t -> Tezt_wrapper.JSON.t) -> unit

Update the configuration file of an sc node. If the sc node is already running, it needs to be restarted manually.

Example: Node.Config_file.update node (JSON.put ("p2p", new_p2p_config))