Module Tezt_tezos.Tez

Helpers for dealing with units of tez.

Please note that none of the functions here perform any bounds checks.

type t

A unit of tez

val of_int : int -> t

Make t from the whole number of tez. This doesn't perform any bounds checks.

val of_mutez_int : int -> t

Make t from the whole number of micro tez. This doesn't perform any bounds checks.

val zero : t

0 tez

val one : t

1 tez

val to_string : t -> string

Convert t to a string.

val to_float : t -> float

Convert t to a float of tez.

val (+) : t -> t -> t

Addition. This doesn't perform any bounds checks.