Module Tezt_tezos.Tezos_regression

Tezos-specific extension for the Regression module.

val replace_variables : string -> string

replace_variables log returns log with all occurrences of variables such as contract addresses, hashes etc. that may change between different runs replaced by constants.

val hooks_custom : ?scrubbed_global_options:string list -> ?replace_variables:(string -> string) -> unit -> Tezt_wrapper.Process.hooks

Custom Tezos-specific regression hooks that replaces Tezos-specific values with constants.

scrubbed_global_options list of global options to scrub. Default: remove global options --base-dir, -d, --endpoint, -E, and --sources.

replace_variables is applied to the output and to the argument of each global option not in scrubbed_global_options. Defaults to replace_variables that replaces key hashes, public keys, contract hashes, and timestamps.

val hooks : Tezt_wrapper.Process.hooks

Hooks that replaces Tezos-specific values with constants.

This is hooks_custom with the default arguments.

val rpc_hooks : RPC_core.rpc_hooks

PRC_hooks that replaces Tezos-specific values with constants and calls Regression.capture on request and response.