Module Teztale.Archiver

type conf = {

name : string;

user : user;

feed : interface list;

}

See parameters of run function for details.

type t = {

process : Tezt_wrapper.Process.t;

conf : conf;

}

val run : ?runner:Tezt_wrapper.Runner.t -> ?path:string -> ?name:string -> node_port:int -> user -> interface list -> t Lwt.t

run ?runner ?path ?name ~node_port user feed

Spawn a teztale archiver with some given parameters:

runner: runner used to spawn the process

path: path of the teztale server executable

name: name the the server used in logs

node_port: port used for the octez node RPCs

user: login and password used by the archiver. Make sure that login and password used have been allowed in the server. See ?users argument of Server.run or Server.add_user.

feed: list of interfaces for teztale servers to feed. If the archiver is feeding multiple servers, every server must accept login info defined by user parameter.