Module Tezt_tezos__Accuser

Spawn Tezos accuser and control them.

type t

Tezos accuser states.

type event = {

name : string;

value : Tezt.JSON.t;

}

Raw events.

val create : protocol:Tezt_tezos.Protocol.t -> ?⁠name:string -> ?⁠color:Tezt.Log.Color.t -> ?⁠event_pipe:string -> ?⁠base_dir:string -> ?⁠runner:Tezt.Runner.t -> Tezt_tezos.Node.t -> t

Create an accuser.

This function just creates the t value, it does not call run.

The path to accuser binary is chosen from the protocol.

The standard output and standard error output of the accuser will be logged with prefix name and color color.

Default event_pipe is a temporary file whose name is derived from name. It will be created as a named pipe so that accuser events can be received.

base_dir corresponds to the (useless) "--base-dir" argument of the tezos-accuser command.

The Node.t parameter is the accuser's node target. The accuser will be configured to be synchronised with the given node, and will communicate with it.

If runner is specified, the accuser will be spawned on this runner using SSH.

val name : t -> string

Get the name of an accuser.

val node_rpc_port : t -> int

Get the RPC port of the associated node.

val terminate : t -> unit Lwt.t

Send SIGTERM to an accuser and wait for it to terminate.

Commands

val run : t -> unit Lwt.t

Spawn tezos-accuser run.

The resulting promise is fulfilled as soon as the accuser has been spawned. It continues running in the background.

Events

exception Terminated_before_event of {

daemon : string;

event : string;

where : string option;

}

Exception raised by wait_for functions if the node terminates before the event.

You may catch or let it propagate to cause the test to fail. daemon is the name of the accuser. event is the name of the event. where is an additional optional constraint, such as "level >= 10".

val wait_for_ready : t -> unit Lwt.t

Wait until the accuser is ready.

More precisely, wait until the node on which the accuser is connected to is bootstrapped, and then, the accuser is ready.

val wait_for : ?⁠where:string -> t -> string -> (Tezt.JSON.t -> 'a option) -> 'a Lwt.t

Wait for a custom event to occur.

Usage: wait_for accuser name filter

If an event named name occurs, apply filter to its value. If filter returns None, continue waiting. If filter returns Some x, return x.

where is used as the where field of the Terminated_before_event exception if the accuser terminates. It should describe the constraint that filter applies, such as "field level exists".

It is advised to register such event handlers before starting the accuser, as if they occur before being registered, they will not trigger your handler. For instance, you can define a promise with let x_event = wait_for accuser "x" (fun x -> Some x) and bind it later with let* x = x_event.

val on_event : t -> (event -> unit) -> unit

Add a callback to be called whenever the accuser emits an event.

Contrary to wait_for functions, this callback is never removed.

Listening to events with on_event will not prevent wait_for promises to be fulfilled. You can also have multiple on_event handlers, although the order in which they trigger is unspecified.

val log_events : t -> unit

Register an event handler that logs all events.

Use this when you need to debug or reverse engineer incoming events. Usually you do not want to keep that in the final versions of your tests.

High-Level Functions

val init : protocol:Tezt_tezos.Protocol.t -> ?⁠name:string -> ?⁠color:Tezt.Log.Color.t -> ?⁠event_pipe:string -> ?⁠base_dir:string -> ?⁠runner:Tezt.Runner.t -> Tezt_tezos.Node.t -> t Lwt.t

Initialize an accuser.

This creates an accuser and run, then waits for the accuser to be ready, and finally returns the accuser.

As the accuser usually needs to be connected to a node, we first wait for the node to be ready and then, run the accuser. If one does not want to wait for the node to be ready, it is necessary to use create and then run.

The path to accuser binary is chosen from the protocol.

The standard output and standard error output of the accuser will be logged with prefix name and color color.

Default event_pipe is a temporary file whose name is derived from name. It will be created as a named pipe so that accuser events can be received.

base_dir corresponds to the (useless) "--base-dir" argument of the tezos-accuser command.

The Node.t parameter is the accuser's node target. The accuser will be configured to be synchronised with the given node, and will communicate with it.

If runner is specified, the accuser will be spawned on this runner using SSH.

val restart : t -> unit Lwt.t

Restart an accuser.

This terminates an accuser, then runs it again and waits for it to be ready.