Module Tezt_tezos__Client

Run Tezos client commands.

type mode =

| Client of Tezt_tezos.Node.t option

| Mockup

| Light of float * Tezt_tezos.Node.t list

| Proxy of Tezt_tezos.Node.t

Mode of the client

type mockup_sync_mode =

| Asynchronous

| Synchronous

The synchronization mode of the client.

Asynchronous mode is when transfer doesn't bake the block.

Synchronous is the default mode (no flag passed to create mockup).

type normalize_mode =

| Readable

| Optimized

| Optimized_legacy

The mode argument of the client's 'normalize data' command

type t

Tezos client states.

val base_dir : t -> string

Get the base directory of a client.

The base directory is the location where clients store their configuration files. It corresponds to the --base-dir option.

val create : ?⁠path:string -> ?⁠admin_path:string -> ?⁠name:string -> ?⁠color:Tezt.Log.Color.t -> ?⁠base_dir:string -> ?⁠node:Tezt_tezos.Node.t -> unit -> t

Create a client.

The standard output and standard error output of the node will be logged with prefix name and color color.

Default base_dir is a temporary directory which is always the same for each name.

The node argument is used to know which port the client should connect to. This node can be overridden for each command, as a client is not actually tied to a node. Most commands require a node to be specified (either with create or with the command itself).

val create_with_mode : ?⁠path:string -> ?⁠admin_path:string -> ?⁠name:string -> ?⁠color:Tezt.Log.Color.t -> ?⁠base_dir:string -> mode -> t

Create a client like create but do not assume Client as the mode.

val get_mode : t -> mode

Get a client's mode. Used with set_mode to temporarily change a client's mode

val set_mode : mode -> t -> unit

Change the client's mode. This function is required for example because we wanna keep a client's wallet. This is impossible if we created a new client from scratch.

RPC calls

type path = string list

Paths for RPCs.

For instance, ["chains"; "main"; "blocks"; "head"] denotes /chains/main/blocks/head.

type query_string = (string * string) list

Query strings for RPCs.

For instance, ["key1", "value1"; "key2", "value2"] denotes ?key1=value1&key2=value2.

type meth =

| GET

| PUT

| POST

| PATCH

HTTP methods for RPCs.

val rpc_path_query_to_string : ?⁠query_string:query_string -> path -> string

rpc_path_query_to_string ["key1", "value1"; "key2", "value2")] ["seg1"; "seg2"] returns /seg1/seg2?key1=value1&key2=value2 where seg1, seg2, key1, key2, value1, and value2 have been appropriately encoded

val rpc : ?⁠node:Tezt_tezos.Node.t -> ?⁠hooks:Tezt.Process.hooks -> ?⁠env:string Tezt.Base.String_map.t -> ?⁠data:Tezt.JSON.u -> ?⁠query_string:query_string -> meth -> path -> t -> Tezt.JSON.t Lwt.t

Use the client to call an RPC.

Run rpc meth path?query_string with data. Fail the test if the RPC call failed. env can be used to customize environment variables, e.g. ("TEZOS_LOG", Protocol.daemon_name protocol ^ ".proxy_rpc->debug") to enable logging.

val spawn_rpc : ?⁠node:Tezt_tezos.Node.t -> ?⁠hooks:Tezt.Process.hooks -> ?⁠env:string Tezt.Base.String_map.t -> ?⁠data:Tezt.JSON.u -> ?⁠query_string:query_string -> meth -> path -> t -> Tezt.Process.t

Same as rpc, but do not wait for the process to exit.

val rpc_list : ?⁠node:Tezt_tezos.Node.t -> t -> string Lwt.t

Run tezos-client rpc list.

val spawn_rpc_list : ?⁠node:Tezt_tezos.Node.t -> t -> Tezt.Process.t

Same as rpc_list, but do not wait for the process to exit.

val shell_header : ?⁠node:Tezt_tezos.Node.t -> ?⁠chain:string -> ?⁠block:string -> t -> string Lwt.t

Run tezos-client rpc /chains/<chain>/blocks/<block>/header/shell.

val spawn_shell_header : ?⁠node:Tezt_tezos.Node.t -> ?⁠chain:string -> ?⁠block:string -> t -> Tezt.Process.t

Same as shell_header, but do not wait for the process to exit.

Admin Client Commands

module Admin : sig ... end

Regular Client Commands

val import_secret_key : ?⁠node:Tezt_tezos.Node.t -> t -> Tezt_tezos.Constant.key -> unit Lwt.t

Run tezos-client import secret key.

val spawn_import_secret_key : ?⁠node:Tezt_tezos.Node.t -> t -> Tezt_tezos.Constant.key -> Tezt.Process.t

Same as import_secret_key, but do not wait for the process to exit.

val activate_protocol : ?⁠node:Tezt_tezos.Node.t -> protocol:Tezt_tezos.Protocol.t -> ?⁠fitness:int -> ?⁠key:string -> ?⁠timestamp:string -> ?⁠timestamp_delay:float -> ?⁠parameter_file:string -> t -> unit Lwt.t

Run tezos-client activate protocol.

If timestamp is not specified explicitely, it is set to now -. timestamp_delay. Default value for timestamp_delay is 365 days, which allows to bake plenty of blocks before their timestamp reach the present (at which point one would have to wait between each block so that peers do not reject them for being in the future).

val spawn_activate_protocol : ?⁠node:Tezt_tezos.Node.t -> protocol:Tezt_tezos.Protocol.t -> ?⁠fitness:int -> ?⁠key:string -> ?⁠timestamp:string -> ?⁠timestamp_delay:float -> ?⁠parameter_file:string -> t -> Tezt.Process.t

Same as activate_protocol, but do not wait for the process to exit.

val bake_for : ?⁠node:Tezt_tezos.Node.t -> ?⁠protocol:Tezt_tezos.Protocol.t -> ?⁠key:string -> ?⁠minimal_timestamp:bool -> ?⁠mempool:string -> ?⁠force:bool -> ?⁠context_path:string -> t -> unit Lwt.t

Run tezos-client bake for.

Default key is Constant.bootstrap1.alias.

val spawn_bake_for : ?⁠node:Tezt_tezos.Node.t -> ?⁠protocol:Tezt_tezos.Protocol.t -> ?⁠key:string -> ?⁠minimal_timestamp:bool -> ?⁠mempool:string -> ?⁠force:bool -> ?⁠context_path:string -> t -> Tezt.Process.t

Same as bake_for, but do not wait for the process to exit.

val show_address : ?⁠show_secret:bool -> alias:string -> t -> Tezt_tezos.Account.key Lwt.t

Run tezos-client show address.

val spawn_show_address : ?⁠show_secret:bool -> alias:string -> t -> Tezt.Process.t

Same as show_address, but do not wait for the process to exit.

val gen_and_show_keys : alias:string -> t -> Tezt_tezos.Account.key Lwt.t

A helper to run tezos-client gen keys followed by tezos-client show address to get the generated key.

val endorse_for : ?⁠node:Tezt_tezos.Node.t -> ?⁠key:string -> t -> unit Lwt.t

Run tezos-client endorse for.

Default key is Constant.bootstrap2.alias.

val spawn_endorse_for : ?⁠node:Tezt_tezos.Node.t -> ?⁠key:string -> t -> Tezt.Process.t

Same as endorse_for, but do not wait for the process to exit.

val transfer : ?⁠node:Tezt_tezos.Node.t -> ?⁠wait:string -> ?⁠burn_cap:Tezt_tezos.Tez.t -> ?⁠fee:Tezt_tezos.Tez.t -> ?⁠gas_limit:int -> ?⁠storage_limit:int -> ?⁠arg:string -> amount:Tezt_tezos.Tez.t -> giver:string -> receiver:string -> t -> unit Lwt.t

Run tezos-client transfer amount from giver to receiver.

val spawn_transfer : ?⁠node:Tezt_tezos.Node.t -> ?⁠wait:string -> ?⁠burn_cap:Tezt_tezos.Tez.t -> ?⁠fee:Tezt_tezos.Tez.t -> ?⁠gas_limit:int -> ?⁠storage_limit:int -> ?⁠arg:string -> amount:Tezt_tezos.Tez.t -> giver:string -> receiver:string -> t -> Tezt.Process.t

Same as transfer, but do not wait for the process to exit.

val set_delegate : ?⁠node:Tezt_tezos.Node.t -> ?⁠wait:string -> src:string -> delegate:string -> t -> unit Lwt.t

Run tezos-client set delegate for <src> to <delegate>.

val spawn_set_delegate : ?⁠node:Tezt_tezos.Node.t -> ?⁠wait:string -> src:string -> delegate:string -> t -> Tezt.Process.t

Same as set_delegate, but do not wait for the process to exit.

val withdraw_delegate : ?⁠node:Tezt_tezos.Node.t -> ?⁠wait:string -> src:string -> t -> unit Lwt.t

Run tezos-client withdraw delegate from <src>.

val spawn_withdraw_delegate : ?⁠node:Tezt_tezos.Node.t -> ?⁠wait:string -> src:string -> t -> Tezt.Process.t

Same as withdraw_delegate, but do not wait for the process to exit.

val get_balance_for : ?⁠node:Tezt_tezos.Node.t -> account:string -> t -> float Lwt.t

Run tezos-client get balance for.

val spawn_get_balance_for : ?⁠node:Tezt_tezos.Node.t -> account:string -> t -> Tezt.Process.t

Same as get_balance_for, but do not wait for the process to exit.

val create_mockup : ?⁠sync_mode:mockup_sync_mode -> ?⁠constants:Tezt_tezos.Protocol.constants -> protocol:Tezt_tezos.Protocol.t -> t -> unit Lwt.t

Run tezos-client create mockup.

val spawn_create_mockup : ?⁠sync_mode:mockup_sync_mode -> ?⁠constants:Tezt_tezos.Protocol.constants -> protocol:Tezt_tezos.Protocol.t -> t -> Tezt.Process.t

Same as create_mockup, but do not wait for the process to exit.

val submit_proposals : ?⁠key:string -> ?⁠wait:string -> proto_hash:string -> t -> unit Lwt.t

Run tezos-client submit proposals for.

Default key is Constant.bootstrap1.alias.

val spawn_submit_proposals : ?⁠key:string -> ?⁠wait:string -> proto_hash:string -> t -> Tezt.Process.t

Same as submit_proposals, but do not wait for the process to exit.

type ballot =

| Nay

| Pass

| Yay

val submit_ballot : ?⁠key:string -> ?⁠wait:string -> proto_hash:string -> ballot -> t -> unit Lwt.t

Run tezos-client submit ballot for.

Default key is Constant.bootstrap1.alias.

val spawn_submit_ballot : ?⁠key:string -> ?⁠wait:string -> proto_hash:string -> ballot -> t -> Tezt.Process.t

Same as submit_ballot, but do not wait for the process to exit.

val originate_contract : ?⁠node:Tezt_tezos.Node.t -> ?⁠wait:string -> ?⁠init:string -> ?⁠burn_cap:Tezt_tezos.Tez.t -> alias:string -> amount:Tezt_tezos.Tez.t -> src:string -> prg:string -> t -> string Lwt.t

Run tezos-client originate contract alias transferring amount from src running prg. Returns the originated contract hash

val spawn_originate_contract : ?⁠node:Tezt_tezos.Node.t -> ?⁠wait:string -> ?⁠init:string -> ?⁠burn_cap:Tezt_tezos.Tez.t -> alias:string -> amount:Tezt_tezos.Tez.t -> src:string -> prg:string -> t -> Tezt.Process.t

Same as originate_contract, but do not wait for the process to exit.

val hash_data : ?⁠expect_failure:bool -> ?⁠hooks:Tezt.Process.hooks -> data:string -> typ:string -> t -> (string * string) list Lwt.t

Run tezos-client hash data .. of type ...

Given that the output of tezos-client is:

Raw packed data: 0x050303 Script-expression-ID-Hash: exprvDnoPjyKeR9FSnvwYg5a1v6mDyB6TmnATwWySSP6VmJxrzQb9E Raw Script-expression-ID-Hash: 0xe0978ddc9329cbd84d25fd15a161a7d2e7e555da91e2a335ece8c8bc11ade245 Ledger Blake2b hash: G7iMrYNckCFRujDLFgtj3cDMCnfeSQ2cmhnDtkh9REc4 Raw Sha256 hash: 0x35ef99f7718e7d1f065bae635780f41c0cd201e9ffb3390ba6ef428c2815fa66 Raw Sha512 hash: 0x2c9ca967bf47f6cc76861693379b7397f65e6a1b6e633df28cf02be0b0d18319ae783b4c199fd61115e000a15a5ba8a292a3b1468c2cfe2b3e3a9fa08d419698 Gas remaining: 1039991.350 units remaining

this function returns the list:

("Raw packed data"; "0x050303") ("Script-expression-ID-Hash"; "exprvDnoPjyKeR9FSnvwYg5a1v6mDyB6TmnATwWySSP6VmJxrzQb9E") ("Raw Script-expression-ID-Hash"; "0xe0978ddc9329cbd84d25fd15a161a7d2e7e555da91e2a335ece8c8bc11ade245") ("Ledger Blake2b hash"; "G7iMrYNckCFRujDLFgtj3cDMCnfeSQ2cmhnDtkh9REc4") ("Raw Sha256 hash"; "0x35ef99f7718e7d1f065bae635780f41c0cd201e9ffb3390ba6ef428c2815fa66") ("Raw Sha512 hash"; "0x2c9ca967bf47f6cc76861693379b7397f65e6a1b6e633df28cf02be0b0d18319ae783b4c199fd61115e000a15a5ba8a292a3b1468c2cfe2b3e3a9fa08d419698") ("Gas remaining"; "1039991.350 units remaining")

If some lines cannot be parsed, warnings are emitted in output and the corresponding lines are omitted from the result.

val spawn_hash_data : ?⁠hooks:Tezt.Process.hooks -> data:string -> typ:string -> t -> Tezt.Process.t

Same as hash_data, but do not wait for the process to exit.

val normalize_data : ?⁠mode:normalize_mode -> ?⁠legacy:bool -> data:string -> typ:string -> t -> string Lwt.t

Run tezos-client normalize data .. of type ...

val spawn_normalize_data : ?⁠mode:normalize_mode -> ?⁠legacy:bool -> data:string -> typ:string -> t -> Tezt.Process.t

Same as normalize_data, but do not wait for the process to exit.

val list_mockup_protocols : t -> string list Lwt.t

Run tezos-client list mockup protocols.

val spawn_list_mockup_protocols : t -> Tezt.Process.t

Same as list_mockup_protocols, but do not wait for the process to exit and do not process stdout.

val migrate_mockup : next_protocol:Tezt_tezos.Protocol.t -> t -> unit Lwt.t

Run tezos-client migrate mockup to.

val spawn_migrate_mockup : next_protocol:Tezt_tezos.Protocol.t -> t -> Tezt.Process.t

Same as migrate_mockup, but do not wait for the process to exit.

High-Level Functions

val init : ?⁠path:string -> ?⁠admin_path:string -> ?⁠name:string -> ?⁠color:Tezt.Log.Color.t -> ?⁠base_dir:string -> ?⁠node:Tezt_tezos.Node.t -> unit -> t Lwt.t

Create a client with mode Client and import all secret keys listed in Constant.all_secret_keys.

val init_mockup : ?⁠path:string -> ?⁠admin_path:string -> ?⁠name:string -> ?⁠color:Tezt.Log.Color.t -> ?⁠base_dir:string -> ?⁠sync_mode:mockup_sync_mode -> ?⁠constants:Tezt_tezos.Protocol.constants -> protocol:Tezt_tezos.Protocol.t -> unit -> t Lwt.t

Create a client with mode Mockup and run create mockup.

Contrary to init, this does not import any secret key, because tezos-client create mockup already initializes the mockup with bootstrap keys.

val init_light : ?⁠path:string -> ?⁠admin_path:string -> ?⁠name:string -> ?⁠color:Tezt.Log.Color.t -> ?⁠base_dir:string -> ?⁠nodes:Tezt_tezos.Node.t list -> ?⁠min_agreement:float -> unit -> (t * Tezt_tezos.Node.t list) Lwt.t

Create a client with mode Light. In addition to the client, the list of nodes is returned, as it was potentially created by this call; and the light mode needs tight interaction with the nodes.