Module Tezt_tezos__Cluster

Set up networks involving large amount of nodes.

Batch Node Creation

val create : ?⁠path:string -> ?⁠name:string -> int -> Tezt_tezos.Node.argument list -> Tezt_tezos.Node.t list

Create several nodes.

Usage: create count arguments

Create count nodes (using Node.create) sharing the same arguments. Each node is named name ^ "." ^ i where i is the index of the node in the resulting list, starting from 1.

Topologies

val connect : Tezt_tezos.Node.t list -> Tezt_tezos.Node.t list -> unit

Connect all nodes of a cluster to all nodes of another cluster.

Usage: connect a b

Add all nodes of b as --peer arguments to all nodes of a. This is a generalization of Node.add_peer for lists.

val clique : Tezt_tezos.Node.t list -> unit

Connect all nodes of a cluster together.

For a cluster node1; node2; ...; nodeN, this adds:

node2; ...; nodeN as --peers to node1;

node3; ...; nodeN as --peers to node2;

etc.

val ring : Tezt_tezos.Node.t list -> unit

Connect nodes to form a ring.

For a cluster node1; node2; ...; nodeN, this adds:

node2 as --peer to node1;

node3 as --peer to node2;

...

and finally node1 as --peer to nodeN.

val star : Tezt_tezos.Node.t -> Tezt_tezos.Node.t list -> unit

Connect nodes to form a star.

Usage: star center other_nodes

Add all other_nodes as --peers to center.

Meta-Topologies

val meta_clique : Tezt_tezos.Node.t list list -> unit

Meta-version of clique.

val meta_ring : Tezt_tezos.Node.t list list -> unit

Meta-version of ring.

val meta_star : Tezt_tezos.Node.t list -> Tezt_tezos.Node.t list list -> unit

Meta-version of star.

Running

val start : Tezt_tezos.Node.t list -> unit Lwt.t

Start a cluster of nodes.

This runs Node.identity_generate, Node.config_init and Node.run for all nodes. This is similar to iterating Node.init except that the node are already created.