Module Tezt_tezos__Node

Spawn Tezos nodes and control them.

type history_mode =

| Archive

| Full

| Rolling

History modes for the node.

type argument =

| Network of string
--network

| History_mode of history_mode
--history-mode

| Expected_pow of int
--expected-pow

| Singleprocess
--singleprocess

| Bootstrap_threshold of int
--bootstrap-threshold (deprecated)

| Synchronisation_threshold of int
--synchronisation-threshold

| Connections of int
--connections

| Private_mode
--private-mode

| Peer of string
--peer

Tezos node command-line arguments.

Not all arguments are available here. Some are simply not implemented, and some are handled separately because they need special care. The latter are implemented as optional labeled arguments (e.g. ?net_port and ?data_dir).

type t

Tezos node states.

val create : ?⁠runner:Tezt.Runner.t -> ?⁠path:string -> ?⁠name:string -> ?⁠color:Tezt.Log.Color.t -> ?⁠data_dir:string -> ?⁠event_pipe:string -> ?⁠net_port:int -> ?⁠rpc_host:string -> ?⁠rpc_port:int -> argument list -> t

Create a node.

This function just creates the t value, it does not call identity_generate nor config_init nor run.

The standard output and standard error output of the node will be logged with prefix name and color color.

Default data_dir is a temporary directory which is always the same for each name.

Default event_pipe is a temporary file whose name is derived from name. It will be created as a named pipe so that node events can be received.

Default values for net_port or rpc_port are chosen automatically with values starting from 16384 (configurable with `--starting-port`). They are used by config_init and by functions from the Client module. They are not used by run, so if you do not call config_init or generate the configuration file through some other means, your node will not listen.

The argument list is a list of configuration options that the node should run with. It is passed to the first run of tezos-node config init. It is also passed to all runs of tezos-node run that occur before tezos-node config init. If Expected_pow is given, it is also used as the default value for identity_generate.

If runner is specified, the node will be spawned on this runner using SSH.

val add_argument : t -> argument -> unit

Add an argument to a node as if it was passed to create.

The argument is passed to the next run of tezos-node config init. It is also passed to all runs of tezos-node run that occur before the next tezos-node config init.

val add_peer : t -> t -> unit

Add a --peer argument to a node.

Usage: add_peer node peer

Same as add_argument node (Peer "127.0.0.1:<PORT>") where <PORT> is the P2P port of peer.

val add_peer_with_id : t -> t -> unit Lwt.t

Add a --peer argument to a node.

Usage: add_peer node peer

Same as add_argument node (Peer "127.0.0.1:<PORT>#<ID>") where <PORT> is the P2P port and <ID> is the identity of peer.

val point_and_id : ?⁠from:t -> t -> string Lwt.t

Get the P2P point and id for a node.

Return "<ADDRESS>:<PORT>#<ID>" where <PORT> is the P2P port and <ID> is the identity of the node.

<ADDRESS> is obtained using Runner.address runner with ?from being the runner of from and runner is the runner of the node. In other words it is the address where from can contact node.

val name : t -> string

Get the name of a node.

val net_port : t -> int

Get the network port given as --net-addr to a node.

val rpc_host : t -> string

Get the RPC host given as --rpc-addr to a node.

val rpc_port : t -> int

Get the RPC port given as --rpc-addr to a node.

val data_dir : t -> string

Get the data-dir of a node.

val runner : t -> Tezt.Runner.t option

Get the runner associated to a node.

Return None if the node runs on the local machine.

val check_error : ?⁠exit_code:int -> ?⁠msg:Tezt.Base.rex -> t -> unit Lwt.t

Wait until a node terminates and check its status.

If the node is not running, or if the process returns an exit code which is not exit_code, or if msg does not match the stderr output, fail the test.

If exit_code is not specified, any non-zero code is accepted. If no msg is given, the stderr is ignored.

val wait : t -> Unix.process_status Lwt.t

Wait until a node terminates and return its status. If the node is not running, make the test fail.

val terminate : t -> unit Lwt.t

Send SIGTERM to a node and wait for it to terminate.

Commands

val identity_generate : ?⁠expected_pow:int -> t -> unit Lwt.t

Run tezos-node identity generate.

val spawn_identity_generate : ?⁠expected_pow:int -> t -> Tezt.Process.t

Same as identity_generate, but do not wait for the process to exit.

val show_history_mode : history_mode -> string

Convert an history mode into a string.

The result is suitable to be passed to the node on the command-line.

val config_init : t -> argument list -> unit Lwt.t

Run tezos-node config init.

module Config_file : sig ... end

val spawn_config_init : t -> argument list -> Tezt.Process.t

Same as config_init, but do not wait for the process to exit.

val run : ?⁠on_terminate:(Unix.process_status -> unit) -> ?⁠event_level:string -> t -> argument list -> unit Lwt.t

Spawn tezos-node run.

The resulting promise is fulfilled as soon as the node has been spawned. It continues running in the background.

event_level specifies the verbosity of the file descriptor sink. This must be at least "notice", which is the level of event "node_is_ready.v0", needed for wait_for_ready. Possible values are therefore: "debug", "info", and "notice". Other values are ignored (verbosity then stays at default "notice").

Events

exception Terminated_before_event of {

daemon : string;

event : string;

where : string option;

}

Exception raised by wait_for functions if the node terminates before the event.

You may catch or let it propagate to cause the test to fail. daemon is the name of the node. event is the name of the event. where is an additional optional constraint, such as "level >= 10".

val wait_for_ready : t -> unit Lwt.t

Wait until the node is ready.

More precisely, wait until a node_is_ready event occurs. If such an event already occurred, return immediately.

val wait_for_level : t -> int -> int Lwt.t

Wait for a given chain level.

More precisely, wait until a node_chain_validator with a level greater or equal to the requested level occurs. If such an event already occurred, return immediately.

val wait_for_identity : t -> string Lwt.t

Wait for the node to read its identity.

More precisely, wait until a read_identity event occurs. If such an event already occurred, return immediately.

Return the identity.

val wait_for : ?⁠where:string -> t -> string -> (Tezt.JSON.t -> 'a option) -> 'a Lwt.t

Wait for a custom event to occur.

Usage: wait_for node name filter

If an event named name occurs, apply filter to its value. If filter returns None, continue waiting. If filter returns Some x, return x.

where is used as the where field of the Terminated_before_event exception if the node terminates. It should describe the constraint that filter applies, such as "field level exists".

It is advised to register such event handlers before starting the node, as if they occur before being registered, they will not trigger your handler. For instance, you can define a promise with let x_event = wait_for node "x" (fun x -> Some x) and bind it later with let* x = x_event.

type event = {

name : string;

value : Tezt.JSON.t;

}

Raw events.

val on_event : t -> (event -> unit) -> unit

Add a callback to be called whenever the node emits an event.

Contrary to wait_for functions, this callback is never removed.

Listening to events with on_event will not prevent wait_for promises to be fulfilled. You can also have multiple on_event handlers, although the order in which they trigger is unspecified.

val log_events : t -> unit

Register an event handler that logs all events.

Use this when you need to debug or reverse engineer incoming events. Usually you do not want to keep that in the final versions of your tests.

High-Level Functions

val init : ?⁠runner:Tezt.Runner.t -> ?⁠path:string -> ?⁠name:string -> ?⁠color:Tezt.Log.Color.t -> ?⁠data_dir:string -> ?⁠event_pipe:string -> ?⁠net_port:int -> ?⁠rpc_host:string -> ?⁠rpc_port:int -> ?⁠event_level:string -> argument list -> t Lwt.t

Initialize a node.

This creates a node, runs identity_generate, config_init and run, then waits for the node to be ready, and finally returns the node.

Arguments are passed to config_init. If you specified Expected_pow, it is also passed to identity_generate. Arguments are not passed to run. If you do not wish the arguments to be stored in the configuration file (which will affect future runs too), do not use init.

val restart : t -> argument list -> unit Lwt.t

Restart a node.

This terminates a node, then runs it again and waits for it to be ready.

If you passed arguments such as Expected_pow or Singleprocess to run they are not automatically passed again. You can pass them to restart, or you can pass other values if you want to restart with other parameters.

val send_raw_data : t -> data:string -> unit Lwt.t

send_raw_data node ~data writes ~data using an IP socket on the net port of node.