Module Tezt_tezos__RPC.Delegates

val get_all : ?⁠node:Tezt_tezos.Node.t -> ?⁠hooks:Tezt.Process.hooks -> ?⁠chain:string -> ?⁠block:string -> Tezt_tezos.Client.t -> string list Lwt.t

Call RPC /chain/chain/blocks/block/context/delegates

val spawn_get_all : ?⁠node:Tezt_tezos.Node.t -> ?⁠hooks:Tezt.Process.hooks -> ?⁠chain:string -> ?⁠block:string -> Tezt_tezos.Client.t -> Tezt.Process.t

Same as get_all, but do not wait for the process to exit.

val get : ?⁠node:Tezt_tezos.Node.t -> ?⁠hooks:Tezt.Process.hooks -> ?⁠chain:string -> ?⁠block:string -> pkh:string -> Tezt_tezos.Client.t -> Tezt.JSON.t Lwt.t

Call RPC /chain/chain/blocks/block/context/delegates/pkh

val spawn_get : ?⁠node:Tezt_tezos.Node.t -> ?⁠hooks:Tezt.Process.hooks -> ?⁠chain:string -> ?⁠block:string -> pkh:string -> Tezt_tezos.Client.t -> Tezt.Process.t

Same as get, but do not wait for the process to exit.

val get_balance : ?⁠node:Tezt_tezos.Node.t -> ?⁠hooks:Tezt.Process.hooks -> ?⁠chain:string -> ?⁠block:string -> pkh:string -> Tezt_tezos.Client.t -> Tezt.JSON.t Lwt.t

Call RPC /chain/chain/blocks/block/context/delegates/pkh/balance

val spawn_get_balance : ?⁠node:Tezt_tezos.Node.t -> ?⁠hooks:Tezt.Process.hooks -> ?⁠chain:string -> ?⁠block:string -> pkh:string -> Tezt_tezos.Client.t -> Tezt.Process.t

Same as get_balance, but do not wait for the process to exit.

val get_deactivated : ?⁠node:Tezt_tezos.Node.t -> ?⁠hooks:Tezt.Process.hooks -> ?⁠chain:string -> ?⁠block:string -> pkh:string -> Tezt_tezos.Client.t -> Tezt.JSON.t Lwt.t

Call RPC /chain/chain/blocks/block/context/delegates/pkh/deactivated

val spawn_get_deactivated : ?⁠node:Tezt_tezos.Node.t -> ?⁠hooks:Tezt.Process.hooks -> ?⁠chain:string -> ?⁠block:string -> pkh:string -> Tezt_tezos.Client.t -> Tezt.Process.t

Same as get_deactivated, but do not wait for the process to exit.

val get_delegated_balance : ?⁠node:Tezt_tezos.Node.t -> ?⁠hooks:Tezt.Process.hooks -> ?⁠chain:string -> ?⁠block:string -> pkh:string -> Tezt_tezos.Client.t -> Tezt.JSON.t Lwt.t

Call RPC /chain/chain/blocks/block/context/delegates/pkh/delegated_balance

val spawn_get_delegated_balance : ?⁠node:Tezt_tezos.Node.t -> ?⁠hooks:Tezt.Process.hooks -> ?⁠chain:string -> ?⁠block:string -> pkh:string -> Tezt_tezos.Client.t -> Tezt.Process.t

Same as get_delegated_balance, but do not wait for the process to exit.

val get_delegated_contracts : ?⁠node:Tezt_tezos.Node.t -> ?⁠hooks:Tezt.Process.hooks -> ?⁠chain:string -> ?⁠block:string -> pkh:string -> Tezt_tezos.Client.t -> Tezt.JSON.t Lwt.t

Call RPC /chain/chain/blocks/block/context/delegates/pkh/delegated_contracts

val spawn_get_delegated_contracts : ?⁠node:Tezt_tezos.Node.t -> ?⁠hooks:Tezt.Process.hooks -> ?⁠chain:string -> ?⁠block:string -> pkh:string -> Tezt_tezos.Client.t -> Tezt.Process.t

Same as get_delegated_contracts, but do not wait for the process to exit.

val get_frozen_balance : ?⁠node:Tezt_tezos.Node.t -> ?⁠hooks:Tezt.Process.hooks -> ?⁠chain:string -> ?⁠block:string -> pkh:string -> Tezt_tezos.Client.t -> Tezt.JSON.t Lwt.t

Call RPC /chain/chain/blocks/block/context/delegates/pkh/frozen_balance

val spawn_get_frozen_balance : ?⁠node:Tezt_tezos.Node.t -> ?⁠hooks:Tezt.Process.hooks -> ?⁠chain:string -> ?⁠block:string -> pkh:string -> Tezt_tezos.Client.t -> Tezt.Process.t

Same as get_frozen_balance, but do not wait for the process to exit.

val get_frozen_balance_by_cycle : ?⁠node:Tezt_tezos.Node.t -> ?⁠hooks:Tezt.Process.hooks -> ?⁠chain:string -> ?⁠block:string -> pkh:string -> Tezt_tezos.Client.t -> Tezt.JSON.t Lwt.t

Call RPC /chain/chain/blocks/block/context/delegates/pkh/frozen_balance_by_cycle

val spawn_get_frozen_balance_by_cycle : ?⁠node:Tezt_tezos.Node.t -> ?⁠hooks:Tezt.Process.hooks -> ?⁠chain:string -> ?⁠block:string -> pkh:string -> Tezt_tezos.Client.t -> Tezt.Process.t

Same as get_frozen_balance_by_cycle, but do not wait for the process to exit.

val get_grace_period : ?⁠node:Tezt_tezos.Node.t -> ?⁠hooks:Tezt.Process.hooks -> ?⁠chain:string -> ?⁠block:string -> pkh:string -> Tezt_tezos.Client.t -> Tezt.JSON.t Lwt.t

Call RPC /chain/chain/blocks/block/context/delegates/pkh/grace_period

val spawn_get_grace_period : ?⁠node:Tezt_tezos.Node.t -> ?⁠hooks:Tezt.Process.hooks -> ?⁠chain:string -> ?⁠block:string -> pkh:string -> Tezt_tezos.Client.t -> Tezt.Process.t

Same as get_grace_period, but do not wait for the process to exit.

val get_staking_balance : ?⁠node:Tezt_tezos.Node.t -> ?⁠hooks:Tezt.Process.hooks -> ?⁠chain:string -> ?⁠block:string -> pkh:string -> Tezt_tezos.Client.t -> Tezt.JSON.t Lwt.t

Call RPC /chain/chain/blocks/block/context/delegates/pkh/staking_balance

val spawn_get_staking_balance : ?⁠node:Tezt_tezos.Node.t -> ?⁠hooks:Tezt.Process.hooks -> ?⁠chain:string -> ?⁠block:string -> pkh:string -> Tezt_tezos.Client.t -> Tezt.Process.t

Same as get_staking_balance, but do not wait for the process to exit.

val get_voting_power : ?⁠node:Tezt_tezos.Node.t -> ?⁠hooks:Tezt.Process.hooks -> ?⁠chain:string -> ?⁠block:string -> pkh:string -> Tezt_tezos.Client.t -> Tezt.JSON.t Lwt.t

Call RPC /chain/chain/blocks/block/context/delegates/pkh/voting_power

val spawn_get_voting_power : ?⁠node:Tezt_tezos.Node.t -> ?⁠hooks:Tezt.Process.hooks -> ?⁠chain:string -> ?⁠block:string -> pkh:string -> Tezt_tezos.Client.t -> Tezt.Process.t

Same as get_voting_power, but do not wait for the process to exit.