Module Tezt_tezos_tezt_cloud.Cli

include module type of Tezt.Cli

module Options : sig ... end

module Logs : sig ... end

module Reports : sig ... end

module Commands : sig ... end

module Selecting_tests : sig ... end

val get : ?default:'a -> (string -> 'a option) -> string -> 'a

val get_opt : (string -> 'a option) -> string -> 'a option

val get_bool : ?default:bool -> string -> bool

val get_bool_opt : string -> bool option

val get_int : ?default:int -> string -> int

val get_int_opt : string -> int option

val get_float : ?default:float -> string -> float

val get_float_opt : string -> float option

val get_string : ?default:string -> string -> string

val get_string_opt : string -> string option

val localhost : bool

When localhost is true, the cloud scenario should be run locally. This can be used to test a scenario before deploying it.

val monitoring : bool

When monitoring is true, VMs are monitored with netdata.

val destroy : bool

destroy is true for destroying temporary resources run by a scenario. Otherwise, this has to be done manually.

val project_id : string option

project_id enables to specify a project id that should be used. This option should be set when the scenario should be run for a different project id than the default one.

val keep_alive : bool

When keep_alive is true, the user must press <enter> to terminate the scenario. This can be used for debugging purpose or for inspecting the state of VMs at the end of a scenario.

val vms : int option

vms enables to set the number of vms that will be spawned.

val vm_base_port : int

First port opened on the VM.

val ports_per_vm : int

Number of consecutive ports opened on the VM from the base port.

val website : bool

When website is true (default) a website is up for summarizing various information related to the experiment.

val website_port : int

website_port determines the port at which the website is set. Default is 8080.

val prometheus : bool

When prometheus is true a Prometheus instance is run locally and metrics of the experiments are exported to prometheus. At the end of the test, the database is snapshotted so that it can be imported later on.

val prometheus_export : bool

If true, exports a Promtheus snapshot at the end of the scenario.

val prometheus_snapshot_filename : string option

Specify the name of the prometheus snapshot.

val prometheus_port : int

Specify the port of the Prometheus instance on the host machine.

val prometheus_scrape_interval : int

Specify the scraping interval of Prometheus.

val grafana : bool

val machine_type : string

Enable to specify a machine type. The string must be a machine description compliant with GCP (ex: "n1-standard-2"). See the different type of machines supported here: https://cloud.google.com/compute/docs/general-purpose-machines#c3d_series.

Default is: "n1-standard-2".

val dockerfile_alias : string option

Specify the dockerfile alias to use. If not specified, the image name will be given by the value of the variable `TEZT_CLOUD`.

val max_run_duration : int

Maximum running time of a VM.

val no_max_run_duration : bool

Deactivate the max run duration parameter.

val proxy : bool

When proxy is true, the scenario is run via a VM instead of the host machine.

val tezt_cloud : string option

val dns_domains : string list

DNS entries to use for the proxy container. Used by gcloud command in user authorized zones

val os : string

OS used by the VM.

val octez_release : string option

The tag of the octez release to be used.

val open_telemetry : bool

Activate the Open Telemetry collector.