Module Tezt_tezos_tezt_cloud.Deployement

type t

Type for resources managed by a deployment.

val deploy : configurations:Configuration.t list -> t Lwt.t

deploy ~configurations deploys the expected number of vms. For each vm, we can specify a base_port which is the first port to be opened and ports_per_vm specify the number of opened port from the base_port. The promise returned by this function is fulfilled when all the vms are deployed. Consequently, it can take some times.

val agents : t -> Agent.t list

get_agents t returns the list of agents deployed.

val terminate : ?exn:exn -> t -> unit Lwt.t

terminate ?exn env t should be called to tear down the machine. Do note that this call may or may not destroy the machine depending on what the user has chosen (see Cli.destroy.

val of_agents : Agent.t list -> t