Module Gcloud.DNS

val list_zones : unit -> (string * string) list Lwt.t

list_zones list available zones in google cloud project

val find_zone_for_subdomain : string -> (string * string) option Lwt.t

find_zone_for_subdomain domain finds a zone suitable to add a subdomain into. It relies on gcloud credentials to find authorized zones. returns None in case no zone is suitable, or Some (zone, domain) with domain the parent domain attached to the zone

val create_zone : domain:string -> zone:string -> unit -> unit Lwt.t

val describe : zone:string -> unit -> string Lwt.t

val get_fqdn : zone:string -> name:string -> string Lwt.t

get_fqdn ~zone ~name returns the fully qualified domain name (FQDN) corresponding to the subdomain or hostname name in the GCP zone zone

val get_value : zone:string -> domain:string -> string option Lwt.t

get_value ~zone ~domain returns the value associated to domain in the zone zone. For example, an ip associated to a hostname.

val add_subdomain : zone:string -> name:string -> value:string -> unit Lwt.t

add_subdomain ~zone ~name ~value adds a dns entry for the domain name name, associated to the value value. The value being an ip

val remove_subdomain : zone:string -> name:string -> value:string -> unit Lwt.t

remove_subdomain ~zone ~name ~value removes the dns record associated to name and value value. If follows the gcloud dns record-sets semantics for removing values associated to the same key.