Module Tezt_tezos_tezt_cloud.Path

val terraform_docker_registry : string

Path to the directory containing the terraform script to setup the docker registry.

val terraform_state_bucket : string

Path to the directory containing the terraform script to generate the bucket storing the terraform state.

val terraform_vm : string

Path to the directory containing the terraform script to generate the VMs.

val dockerfile : alias:string -> string

Path to the directory containing dockerfiles for each user.

val zcash_params : string

Path to the zcash_params directory that can be used for the dockerfiles.

val dal_trusted_setup : string

Path to the DAL trusted setup directory that can be used for the dockerfiles.

val default_binaries_path : unit -> string

Default path where images can be found on the docker container.

val self : string

Path to the current binary.

val proxy_deployement : tezt_cloud:string -> string

Path where to find the encoding of tezt cloud deployement.

val grafana_dashboards : string

Path where are stored grafana dashboards.