Module Tezt_tezos_tezt_cloud.Tezt_cloud

module Agent : sig ... end

module Configuration : sig ... end

module Cloud : sig ... end

val register : tags:string list -> unit

register ~tags register a set of jobs that can be used for setting requirements related to cloud scenarios. Some tags can be given for all the registered jobs.