Module Tezt_tezos_tezt_cloud.Web

type t

val start : agents:Agent.t Stdlib.List.t -> t Lwt.t

start ~agents env starts a webpage with experimentations information if Cli.website is true.

val shutdown : t -> unit Lwt.t

shutdown website shutdowns the website.

val push_metric : t -> ?labels:(string * string) list -> name:string -> float -> unit

push_metric ?labels ~name value push a metric into the page /metrics.txt. This page can be scrapped by Prometheus if activated.

val write : t -> agents:Agent.t Stdlib.List.t -> unit Lwt.t

type service = {

name : string;

url : string;

}

val add_service : t -> agents:Agent.t Stdlib.List.t -> service -> unit Lwt.t