Module Client.Admin

Run octez-admin-client commands.

val trust_address : ?endpoint:endpoint -> peer:Node.t -> t -> unit Lwt.t

Ask a node to trust the address and port of another node.

val untrust_address : ?endpoint:endpoint -> peer:Node.t -> t -> unit Lwt.t

Ask a node to untrust the address and port of another node.

val spawn_trust_address : ?endpoint:endpoint -> peer:Node.t -> t -> Tezt_wrapper.Process.t

Same as trust_address, but do not wait for the process to exit.

val spawn_untrust_address : ?endpoint:endpoint -> peer:Node.t -> t -> Tezt_wrapper.Process.t

Same as untrust_address, but do not wait for the process to exit.

val connect_address : ?endpoint:endpoint -> peer:Node.t -> t -> unit Lwt.t

Connect a node to another peer.

val spawn_connect_address : ?endpoint:endpoint -> peer:Node.t -> t -> Tezt_wrapper.Process.t

Same as connect_address, but do not wait for the process to exit.

val kick_peer : ?endpoint:endpoint -> peer:string -> t -> unit Lwt.t

Kick a peer.

peer is the identity of the peer to kick. You can get it with Node.wait_for_identity for instance.

val spawn_kick_peer : ?endpoint:endpoint -> peer:string -> t -> Tezt_wrapper.Process.t

Same as kick_peer, but do not wait for the process to exit.

val ban_peer : ?endpoint:endpoint -> peer:string -> t -> unit Lwt.t

Ban a peer.

peer is the identity of the peer to ban. You can get it with Node.wait_for_identity for instance.

val spawn_ban_peer : ?endpoint:endpoint -> peer:string -> t -> Tezt_wrapper.Process.t

Same as ban_peer, but do not wait for the process to exit.

val p2p_stat : ?endpoint:endpoint -> t -> string Lwt.t

Run octez-admin-client p2p stat.

val spawn_p2p_stat : ?endpoint:endpoint -> t -> Tezt_wrapper.Process.t

Same as p2p_stat, but do not wait for the process to exit.

val inject_protocol : ?endpoint:endpoint -> protocol_path:string -> t -> string Lwt.t

Run octez-admin-client inject protocol <protocol_path>.

Returns the hash of the injected protocol.

val spawn_inject_protocol : ?endpoint:endpoint -> protocol_path:string -> t -> Tezt_wrapper.Process.t

Same as inject_protocol, but do not wait for the process to exit.

val list_protocols : ?endpoint:endpoint -> t -> string list Lwt.t

Run octez-admin-client list protocols and return the list of protocol hashes.

val spawn_list_protocols : ?endpoint:endpoint -> t -> Tezt_wrapper.Process.t

Same as list_protocols, but do not wait for the process to exit.

val protocol_environment : ?endpoint:endpoint -> t -> string -> string Lwt.t

Run octez-admin-client protocol environment on a protocol hash.

Return its environment version as a string such as "V1".

val spawn_protocol_environment : ?endpoint:endpoint -> t -> string -> Tezt_wrapper.Process.t

Same as protocol_environment, but do not wait for the process to exit.