Module Tezt_tezos.Codec

Run octez-codec commands.

In all functions below, path can be used to override the path to octez-codec, which is Constant.tezos_codec by default.

val encode : ?path:string -> ?hooks:Tezt_wrapper.Process.hooks -> name:string -> Tezt_wrapper.JSON.u -> string Lwt.t

Run octez-codec encode <name> from <json>.

val spawn_encode : ?path:string -> ?hooks:Tezt_wrapper.Process.hooks -> name:string -> Tezt_wrapper.JSON.u -> Tezt_wrapper.Process.t

Same as encode, but do not wait for the process to exit.

val decode : ?path:string -> ?hooks:Tezt_wrapper.Process.hooks -> name:string -> string -> Tezt_wrapper.JSON.t Lwt.t

Run octez-codec decode <name> from <binary>.

val spawn_decode : ?path:string -> ?hooks:Tezt_wrapper.Process.hooks -> name:string -> string -> Tezt_wrapper.Process.t

Same as decode, but do not wait for the process to exit.

val dump_encodings : ?path:string -> unit -> Tezt_wrapper.JSON.t Lwt.t

Run octez-codec dump encodings.

val spawn_dump_encodings : ?path:string -> unit -> Tezt_wrapper.Process.t

Same as dump_encodings, but do not wait for the process to exit.

val describe_binary_schema : ?path:string -> id:string -> unit -> string Lwt.t

val describe_json_schema : ?path:string -> id:string -> unit -> string Lwt.t