Module Tezt_tezos.Protocol

type t =

| Edo

| Florence

| Alpha

Protocols we may want to test with.

type constants =

| Constants_sandbox

| Constants_mainnet

| Constants_test

Protocol parameters.

These values denote which file to use from the "parameters" directory.

val default_constants : constants

The default constants used by tests: Constants_sandbox.

val name : t -> string

Get the name of a protocol, capitalized (e.g. "Edo").

val tag : t -> string

Get the name of a protocol as a tag, for use when registering tests (e.g. "edo").

val hash : t -> string

Get the full hash of a protocol.

val parameter_file : ?⁠constants:constants -> t -> string

Get the location of the parameter file.

This returns the path to one of the parameter files of the "parameters" directory of the protocol, relative to the root of the repository.

val accuser : t -> string

Get the path of the accuser of a protocol, such as "./tezos-accuser-alpha".

val daemon_name : t -> string

Get the part of the daemon name that is specific to a protocol (e.g. "008-PtEdo2Zk").

val encoding_prefix : t -> string

Get the part which is added at the beginning of all encoding names.

It turns out this is equal to what the daemon_name function returns.

type parameter_overrides = (string list * string option) list

Values to override in protocol parameters.

The are pairs of JSON paths and optional values that can be used to override or remove (when the value is None) the default parameters when activating protocol.

val write_parameter_file : protocol:t -> parameter_overrides -> string Lwt.t

Write a protocol parameter file.

Default values are used, except for values given as parameter_overrides.

val next_protocol : t -> t option

Get the successor of a protocol.

WARNING: use of this function is discouraged, because:

a protocol may have several possible successors;

it prevents the type-checker from telling you that your test can no longer run when removing a protocol.

val previous_protocol : t -> t option

Get the predecessor of a protocol.

WARNING: use of this function is discouraged, because it prevents the type-checker from telling you that your test can no longer run when removing a protocol.

val all : t list

Get the list of all protocols.

WARNING: use of this function is discouraged, because:

if we add protocols such as demo-noop or demo-counter to type t, it would be awkward to add them to all if all only contained "real" protocols up to now, and it would also be awkward not to add them to all;

it is easier for maintainers if the list of protocols a test runs on is clearly defined where the test is registered.

val current_mainnet : t

The protocol that is currently active on Mainnet.

Use this when registering tests that are not expected to break in future protocols.

val register_test : __FILE__:string -> title:string -> tags:string list -> (t -> unit Lwt.t) -> protocols:t list -> unit

Register a test that uses the protocol.

This is the same as Test.register except that:

the name of the protocol is automatically added at the beginning of title;

the name of the protocol is automatically added as a tag in tags;

the function that implements the test takes a protocol as parameter;

an additional parameter gives the list of protocols to run the test on.

The list of protocol is the last parameter because this function is intended to be applied partially, without this list; the test function being the parameter of type t -> unit Lwt.t before the list. The list of protocols is specified in main.ml, next to the registration of all other tests, so that it is easy to see at a glance which tests run on which protocols.

If your test involves several protocols (for instance to test migration), use Test.register directly.

val register_regression_test : __FILE__:string -> title:string -> tags:string list -> output_file:string -> ?⁠regression_output_path:string -> (t -> unit Lwt.t) -> protocols:t list -> unit

Register a regression test that uses the protocol.

This is the same as Regression.register, with the same differences as Protocol.register_test compared to Test.register.